Feat : LevelDirectUpdateProcessor 내부 공통 로직 모듈화
